refactor: rename `_documentation.py` to `_docstring.py`
