+++ src/library_analyzer/processing/api/model/_docstring.py
